+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -205,6 +205,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The ray traced screen space shadow history for directional, spot and point lights is discarded if the light transform has changed.
 - Changed the behavior for ray tracing in case a mesh renderer has both transparent and opaque submeshes.
 - Improve history buffer management
+- Remove redundant FrameSettings RealTimePlanarReflection
+- Improved a bit the GC calls generated during the rendering.
 
 ## [7.1.1] - 2019-09-05